Added :edit and :update of Item

+++ db/migrate/20210511102016_create_questions.rb
@@ -4,7 +4,7 @@ class CreateQuestions < ActiveRecord::Migration[6.1]
       t.string :title
       t.text :body
       t.string :answer
-      t.int :payment
+      t.float :payment
 
       t.timestamps
     end
